# specs/openapi/dist/api-bundle.yaml
openapi: 3.1.0
info:
  title: Virtual Companion API
  version: 0.1.0
  description: Contract-first HTTP API (ADR-0002). This document is the single source of truth; Java interfaces and TypeScript
    clients are generated from it and must not be hand-edited or used to back-drive the contract. The ErrorEnvelope shape
    is uniform across every error response; NOT_FOUND_OR_FORBIDDEN hides resource existence so a missing resource and a forbidden
    one are indistinguishable to the caller.
security:
- bearerAuth: []
paths:
...
  /api/v1/relationships/{relationshipId}/memories/candidates:
    post:
      operationId: createMemoryCandidate
      summary: Create a memory candidate from model extraction (PENDING_CONFIRMATION)
      description: The model-extraction entry point that persists a memory candidate (vc.create_memory_candidate). The candidate
        is always created in PENDING_CONFIRMATION and never ACCEPTED; canonical memory is reached only by the user confirming
        it later (INV-MEM-001/002). Alpha scope is SESSION or RELATIONSHIP; SESSION additionally requires a conversation binding.
        The caller identity is the owner. Requires a server-verified authentication context; a foreign relationship maps to
        NOT_FOUND_OR_FORBIDDEN.
      parameters:
      - $ref: '#/components/parameters/RelationshipId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCandidateCreateRequest'
      responses:
        '200':
          description: The created candidate, in PENDING_CONFIRMATION status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/relationships/{relationshipId}/memories:
    get:
      operationId: listMemories
      summary: List the caller's memory for a relationship
      description: Returns the memory candidates and canonical records scoped to a relationship (vc.list_memory). Soft-deleted
        records are excluded unless includeDeleted is set. Another relationship's memory is never disclosed; a foreign relationship
        yields no rows and never discloses existence. Requires a server-verified authentication context.
      parameters:
      - $ref: '#/components/parameters/RelationshipId'
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
        description: Include soft-deleted memory in the result (default false).
      responses:
        '200':
          description: Memory scoped to the relationship (excluding deleted by default)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/memories/{memoryId}:
    get:
      operationId: getMemory
      summary: Fetch a memory owned by the caller
      description: Returns a non-deleted memory owned by the caller (vc.get_memory). A foreign, absent or deleted id maps
        to NOT_FOUND_OR_FORBIDDEN; existence is never disclosed. Requires a server-verified authentication context.
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: The caller's memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      operationId: updateMemory
      summary: Edit a memory's content (status-preserving, idempotent)
      description: 'Edits the summary of an owned, non-deleted, editable memory (vc.update_memory). Only PENDING_CONFIRMATION
        and ACCEPTED memories are editable; status is never changed, so this cannot create or revive a canonical record. Idempotent:
        applying the same summary twice succeeds both times. A foreign, absent, deleted or dead-end id maps to NOT_FOUND_OR_FORBIDDEN.
        Requires a server-verified authentication context.'
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryUpdateRequest'
      responses:
        '200':
          description: The updated memory, status unchanged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteMemory
      summary: Soft-delete a memory (idempotent)
      description: 'Soft-deletes an owned memory (vc.delete_memory). Idempotent on already- deleted rows: a duplicate delete
        of the caller''s own memory returns success. A foreign or absent id maps to NOT_FOUND_OR_FORBIDDEN; existence is never
        disclosed. Requires a server-verified authentication context.'
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: The memory was soft-deleted (or was already deleted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/memories/{memoryId}/confirm:
    post:
      operationId: confirmMemoryCandidate
      summary: Confirm a pending candidate into canonical memory
      description: 'The sole path to a canonical (ACCEPTED) record (vc.confirm_memory_candidate). Only a PENDING_CONFIRMATION
        candidate can be confirmed; confirming an already-confirmed or otherwise transitioned memory is rejected and maps
        to NOT_FOUND_OR_FORBIDDEN (confirm is intentionally not idempotent into canonical). R44 (V68 / §11.11): an optional
        supersedeMemoryId marks the referenced active canonical memory SUPERSEDED in the same transaction — an explicit user
        choice, never automatic conflict detection. Requires a server-verified authentication context.'
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryConfirmRequest'
      responses:
        '200':
          description: The memory, now in ACCEPTED status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '400':
          description: The supersede target is not an active canonical memory of this relationship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/memories/{memoryId}/reject:
    post:
      operationId: rejectMemoryCandidate
      summary: Reject a pending candidate
      description: Rejects a PENDING_CONFIRMATION candidate (vc.reject_memory_candidate). Only a pending candidate can be
        rejected; rejecting an already-transitioned memory is rejected and maps to NOT_FOUND_OR_FORBIDDEN. Requires a server-verified
        authentication context.
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: The memory, now in REJECTED status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/memories/{memoryId}/evidence:
    get:
      operationId: listMemoryEvidence
      summary: List the source Evidence of a memory
      description: Returns the source Evidence chain of an owned, non-deleted memory (vc.list_memory_evidence). A foreign,
        absent or deleted id yields no rows, indistinguishable from a real memory that carries no evidence, so existence is
        never disclosed. Requires a server-verified authentication context.
      parameters:
      - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: The ordered source Evidence for the memory (possibly empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemoryEvidence'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/memories/auto-save:
    get:
      operationId: getMemoryAutoSave
      summary: The caller's memory auto-save switch (MEM-AUTO-SAVE, §7.4)
      description: The per-owner deterministic low-sensitivity auto-save switch (vc.get_memory_auto_save_pref, V66). Default
        true — §7.4 lists auto-save as the controlled Beta baseline, limited to the fixed low-sensitivity whitelist categories;
        the owner can turn it off at any time (可随时撤销). Requires a server-verified authentication context.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: The current switch state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryAutoSavePref'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: setMemoryAutoSave
      summary: Flip the memory auto-save switch (MEM-AUTO-SAVE, §7.4)
      description: Flips the per-owner auto-save switch (vc.set_memory_auto_save_pref, V66). Turning it off stops every future
        auto-save immediately; already auto-saved rows stay (each is individually deletable). Requires a server-verified authentication
        context.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryAutoSaveRequest'
      responses:
        '200':
          description: The stored switch state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryAutoSavePref'
        '400':
          description: A missing or invalid enabled value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorEnvelope:
      type: object
      description: Unified error shape. NOT_FOUND_OR_FORBIDDEN never discloses whether a resource exists; every error response
        uses this envelope.
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Stable, non-sensitive human-readable explanation
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (never sensitive data)
...
    Memory:
      type: object
      description: A memory candidate or canonical record owned by exactly one user within a relationship. Model extraction
        only creates PENDING_CONFIRMATION candidates; ACCEPTED (canonical) is reached by user confirmation (INV-MEM-001/002)
        or, since V66, the deterministic low-sensitivity auto-save rule — such rows carry autoSaved=true so the UI can mark
        them (界面明示). status follows the memory-candidate-statuses catalog.
      required:
      - memoryId
      - scope
      - summary
      - status
      - autoSaved
      properties:
        memoryId:
          type: string
          description: Stable owner-scoped memory identifier
        scope:
          type: string
          description: memory-scopes catalog code (SESSION or RELATIONSHIP in Alpha)
        summary:
          type: string
          description: Memory content summary
        status:
          type: string
          description: Current memory-candidate-statuses catalog code
        conversationId:
          type: string
          description: Owning conversation identifier, present for SESSION-scoped memory
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (RFC 3339)
        deletedAt:
          type: string
          format: date-time
          description: Soft-deletion timestamp; present only when includeDeleted is true
        autoSaved:
          type: boolean
          description: True only for rows saved by the deterministic low-sensitivity auto rule (V66).
        supersededAt:
          type: string
          format: date-time
          description: R44 (V68) explicit-supersede timestamp; present only when this canonical memory was replaced by a newer
            confirmed one (§11.11). Like deletedAt it is a tombstone column, not a status value.
        supersededByMemoryId:
          type: string
          description: The replacing memory identifier, present only with supersededAt.
        eventAt:
          type: string
          format: date-time
          description: R44 (V68 / §11.12) event anchor; present only on event memories (any event field requires eventAt).
        eventStatus:
          type: string
          description: memory-event-statuses catalog code (PLANNED / IN_PROGRESS / COMPLETED / CANCELLED / UNKNOWN).
        eventExpiresAt:
          type: string
          format: date-time
          description: Event expiry; once passed, reads lazily move the row to EXPIRED (§11.12 Event 到期, V61-style lazy sweep).
    MemoryAutoSavePref:
      type: object
      description: The per-owner memory auto-save switch state (V66, §7.4).
      required:
      - enabled
      properties:
        enabled:
          type: boolean
          description: Whether the deterministic low-sensitivity auto-save rule is active for the owner.
    MemoryAutoSaveRequest:
      type: object
      description: Body for flipping the memory auto-save switch.
      required:
      - enabled
      properties:
        enabled:
          type: boolean
          description: The desired switch state.
    MemoryCandidateCreateRequest:
      type: object
      description: Body for creating a memory candidate from model extraction.
      required:
      - scope
      - summary
      properties:
        scope:
          type: string
          description: memory-scopes catalog code; SESSION or RELATIONSHIP in Alpha
        summary:
          type: string
          description: Memory content summary proposed by the model
        conversationId:
          type: string
          description: Required when scope is SESSION (the binding conversation)
        evidence:
          type: array
          description: Cited source references supporting the candidate
          items:
            type: string
        eventAt:
          type: string
          format: date-time
          description: 'R44 (V68 / §11.12): optional event anchor; any event field requires it.'
        eventStatus:
          type: string
          description: memory-event-statuses catalog code; defaults to PLANNED when eventAt is set.
        eventExpiresAt:
          type: string
          format: date-time
          description: Optional event expiry; must be strictly after eventAt.
    MemoryUpdateRequest:
      type: object
      description: Body for editing a memory's content (status-preserving).
      required:
      - summary
      properties:
        summary:
          type: string
          description: The new memory content summary
        eventAt:
          type: string
          format: date-time
          description: 'R44: optional event edit; absent leaves the stored value unchanged.'
        eventStatus:
          type: string
          description: 'R44: optional memory-event-statuses code edit; absent leaves the stored value unchanged.'
        eventExpiresAt:
          type: string
          format: date-time
          description: 'R44: optional event expiry edit; absent leaves the stored value unchanged.'
    MemoryConfirmRequest:
      type: object
      description: R44 (V68 / §11.11) optional confirm body — the explicit supersede choice. The target must be an active
        canonical memory of the same relationship; a stale or invalid target is 400 INVALID_REQUEST (the id came from the
        caller's own list, so it is not an existence-hidden case). Absent body or field means a plain confirm.
      properties:
        supersedeMemoryId:
          type: string
          description: Identifier of the canonical memory this confirmation replaces.
    MemoryEvidence:
      type: object
      description: A cited source reference supporting a memory candidate.
      required:
      - evidenceId
      - sourceRef
      properties:
        evidenceId:
          type: string
          description: Stable owner-scoped evidence identifier
        sourceRef:
          type: string
          description: Cited source reference
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (RFC 3339)
...
  parameters:
...
    RelationshipId:
      name: relationshipId
      in: path
      required: true
      schema:
        type: string
      description: Owner-scoped relationship identifier
...
    MemoryId:
      name: memoryId
      in: path
      required: true
      schema:
        type: string
      description: Owner-scoped memory identifier
...